+++ geotrek/altimetry/tests/test_commands.py
@@ -9,6 +9,11 @@ from django.test import TransactionTestCase
 from geotrek.altimetry.functions import RasterValue
 from geotrek.altimetry.models import Dem
 
+from geotrek.core.models import Path
+from geotrek.core.tests.factories import PathFactory
+
+from django.contrib.gis.geos import LineString
+
 
 class CommandLoadDemTest(TransactionTestCase):
     """
@@ -16,7 +21,7 @@ class CommandLoadDemTest(TransactionTestCase):
     Use of TransactionTestCase to avoid nesting transaction caused by raster2pgsql generated sql file import.
     """
 
-    def test_success_without_dem(self):
+    def test_success_without_replace(self):
         output_stdout = StringIO()
         filename = os.path.join(os.path.dirname(__file__), 'data', 'elevation.tif')
         call_command('loaddem', filename, verbosity=2, stdout=output_stdout)
@@ -26,7 +31,7 @@ class CommandLoadDemTest(TransactionTestCase):
         value = dems.first()
         self.assertAlmostEqual(value.int, 343.600006103516)
 
-    def test_success_with_dem(self):
+    def test_success_with_replace(self):
         output_stdout = StringIO()
         filename = os.path.join(os.path.dirname(__file__), 'data', 'elevation.tif')
         call_command('loaddem', filename, verbosity=2, stdout=output_stdout)
@@ -38,6 +43,20 @@ class CommandLoadDemTest(TransactionTestCase):
         dems = Dem.objects.all().annotate(int=RasterValue('rast', Point(x=605600, y=6650000, srid=2154)))
         value = dems.first()
         self.assertAlmostEqual(value.int, 343.600006103516)
+
+    def test_success_without_replace_update_altimetry(self):
+        output_stdout = StringIO()
+        filename = os.path.join(os.path.dirname(__file__), 'data', 'elevation.tif')
+        self.path = PathFactory.create(geom=LineString((605600, 6650000), (605900, 6650010), srid=2154))
+        call_command('loaddem', filename, update_altimetry=True, verbosity=2, stdout=output_stdout)
+        self.assertIn('DEM successfully loaded.', output_stdout.getvalue())
+        self.assertIn('Everything looks fine, we can start loading DEM', output_stdout.getvalue())
+        self.assertIn('Updating 3d geometries.', output_stdout.getvalue())
+        dems = Dem.objects.all().annotate(int=RasterValue('rast', Point(x=605600, y=6650000, srid=2154)))
+        value = dems.first()
+        self.assertAlmostEqual(value.int, 343.600006103516)
+        path = Path.objects.get(pk=self.path.pk)
+        self.assertAlmostEqual(path.geom_3d.coords[-1][-1], 188)
 
     def test_fail_table_altimetry_dem(self):
         """ DEM data already exist """